Create README.md
Create a program that allows the user to do some basic functions.

First, ask the user if they would like to find out sqrt, log or factorial of a number, then return the results.

Here is the sample output: 

Welcome to the simple math helper.
What would you like to calculate?
1. Sqrt
2. Log
3. Factorial
> 1
Enter the number to sqrt:
>9
3
To complete this assignment, you will need to do some research. First, check to see if the math functions are part of your language. Next, see if they are included with the language, but need to be imported. Finally, look externally for them.